# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AE Scientist API",
    "description": "Transform LLM conversations into actionable AE ideas",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "Get a paginated list of conversations for the current user.\n\nQuery Parameters:\n- conversation_status: Filter by \"draft\" or \"with_research\" (optional)\n- run_status: Filter by \"pending\", \"running\", \"completed\", or \"failed\" (optional)",
        "operationId": "list_conversations_api_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "conversation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Status"
            }
          },
          {
            "name": "run_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/idea": {
      "get": {
        "summary": "Get Idea",
        "description": "Get the idea for a conversation.",
        "operationId": "get_idea_api_conversations__conversation_id__idea_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IdeaGetResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Idea Api Conversations  Conversation Id  Idea Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Idea",
        "description": "Manually update an idea with all fields.",
        "operationId": "update_idea_api_conversations__conversation_id__idea_patch",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaRefinementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IdeaUpdateResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Update Idea Api Conversations  Conversation Id  Idea Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/idea/versions": {
      "get": {
        "summary": "Get Idea Versions",
        "description": "Get all versions of an idea for a conversation.",
        "operationId": "get_idea_versions_api_conversations__conversation_id__idea_versions_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IdeaVersionsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Idea Versions Api Conversations  Conversation Id  Idea Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/idea/versions/{version_id}/activate": {
      "post": {
        "summary": "Activate Idea Version",
        "description": "Recover a previous version by creating a new version with the same content.",
        "operationId": "activate_idea_version_api_conversations__conversation_id__idea_versions__version_id__activate_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IdeaUpdateResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Activate Idea Version Api Conversations  Conversation Id  Idea Versions  Version Id  Activate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "import_date": {
            "type": "string",
            "title": "Import Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "idea_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idea Title"
          },
          "idea_abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idea Abstract"
          },
          "last_user_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last User Message Content"
          },
          "last_assistant_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Assistant Message Content"
          },
          "manual_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Title"
          },
          "manual_hypothesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Hypothesis"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "import_date",
          "created_at",
          "updated_at",
          "user_id",
          "user_name",
          "user_email"
        ],
        "title": "ConversationListItem",
        "description": "Response model for dashboard list view."
      },
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "List of conversations"
          }
        },
        "type": "object",
        "required": [
          "conversations"
        ],
        "title": "ConversationListResponse",
        "description": "Response for listing conversations."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Idea": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id",
            "description": "Idea ID"
          },
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id",
            "description": "Associated conversation ID"
          },
          "active_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdeaVersion"
              },
              {
                "type": "null"
              }
            ],
            "description": "Currently active version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO format creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO format last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "conversation_id",
          "created_at",
          "updated_at"
        ],
        "title": "Idea",
        "description": "Represents a research idea with its active version."
      },
      "IdeaGetResponse": {
        "properties": {
          "idea": {
            "$ref": "#/components/schemas/Idea",
            "description": "Retrieved idea"
          }
        },
        "type": "object",
        "required": [
          "idea"
        ],
        "title": "IdeaGetResponse",
        "description": "Get idea response."
      },
      "IdeaRefinementRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Idea title"
          },
          "short_hypothesis": {
            "type": "string",
            "title": "Short Hypothesis",
            "description": "Short hypothesis of the idea"
          },
          "related_work": {
            "type": "string",
            "title": "Related Work",
            "description": "Related work or background"
          },
          "abstract": {
            "type": "string",
            "title": "Abstract",
            "description": "Abstract of the idea"
          },
          "experiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiments",
            "description": "List of experiments"
          },
          "expected_outcome": {
            "type": "string",
            "title": "Expected Outcome",
            "description": "Expected outcome of the experiments"
          },
          "risk_factors_and_limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Factors And Limitations",
            "description": "Risk factors and limitations"
          }
        },
        "type": "object",
        "required": [
          "title",
          "short_hypothesis",
          "related_work",
          "abstract",
          "experiments",
          "expected_outcome",
          "risk_factors_and_limitations"
        ],
        "title": "IdeaRefinementRequest",
        "description": "Request model for manually updating an idea with all fields."
      },
      "IdeaUpdateResponse": {
        "properties": {
          "idea": {
            "$ref": "#/components/schemas/Idea",
            "description": "Updated idea"
          }
        },
        "type": "object",
        "required": [
          "idea"
        ],
        "title": "IdeaUpdateResponse",
        "description": "Update idea response."
      },
      "IdeaVersion": {
        "properties": {
          "version_id": {
            "type": "integer",
            "title": "Version Id",
            "description": "Version ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Research idea title"
          },
          "short_hypothesis": {
            "type": "string",
            "title": "Short Hypothesis",
            "description": "Short hypothesis statement"
          },
          "related_work": {
            "type": "string",
            "title": "Related Work",
            "description": "Related work and background"
          },
          "abstract": {
            "type": "string",
            "title": "Abstract",
            "description": "Detailed abstract of the idea"
          },
          "experiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiments",
            "description": "List of proposed experiments"
          },
          "expected_outcome": {
            "type": "string",
            "title": "Expected Outcome",
            "description": "Expected outcome of the research"
          },
          "risk_factors_and_limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Factors And Limitations",
            "description": "Risk factors and limitations"
          },
          "is_manual_edit": {
            "type": "boolean",
            "title": "Is Manual Edit",
            "description": "Whether this version was manually edited by user"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number",
            "description": "Version number for ordering"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO format creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "title",
          "short_hypothesis",
          "related_work",
          "abstract",
          "experiments",
          "expected_outcome",
          "risk_factors_and_limitations",
          "is_manual_edit",
          "version_number",
          "created_at"
        ],
        "title": "IdeaVersion",
        "description": "Represents a single version of a research idea."
      },
      "IdeaVersionsResponse": {
        "properties": {
          "versions": {
            "items": {
              "$ref": "#/components/schemas/IdeaVersion"
            },
            "type": "array",
            "title": "Versions",
            "description": "List of idea versions"
          }
        },
        "type": "object",
        "required": [
          "versions"
        ],
        "title": "IdeaVersionsResponse",
        "description": "Get idea versions response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}